+++ .tape.js
@@ -1,5 +1,11 @@
 module.exports = {
 	'basic': {
 		message: 'supports basic usage'
+	},
+	'missing': {
+		message: 'supports missing includes',
+		warnings: {
+			text: 'HTML not found'
+		}
 	}
 };